Refactor: [M3-6278]: AccessTable
