+++ packages/manager/src/features/ObjectStorage/AccessKeyLanding/ViewPermissionsDrawer.tsx
@@ -1,9 +1,9 @@
-import { ObjectStorageKey } from '@linode/api-v4/lib/object-storage';
 import * as React from 'react';
-import Typography from 'src/components/core/Typography';
 import Drawer from 'src/components/Drawer';
+import Typography from 'src/components/core/Typography';
+import { AccessTable } from './AccessTable';
+import { ObjectStorageKey } from '@linode/api-v4/lib/object-storage';
 
-import { AccessTable } from './LimitedAccessControls';
 export interface Props {
   open: boolean;
   onClose: () => void;
